% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cengen_expression}
\alias{cengen_expression}
\alias{cengen_TPM}
\alias{cengen_proportion}
\alias{cengen_UMI}
\alias{cengen_nCells}
\alias{cengen_TPM_bulk}
\alias{cengen_proportion_bulk}
\alias{cengen_UMI_bulk}
\alias{cengen_nCells_bulk}
\title{CeNGEN single-cell expression level.}
\format{
A matrix with one row per gene and one column by single-cell cluster:
\describe{
  \item{row name}{Wormbase ID of the gene}
  \item{column name}{cluster identification, usually corresponds to a neuron type}
  \item{value}{normalized gene expression level}
}

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 130 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 130 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 130 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 130 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 121 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 121 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 121 columns.

An object of class \code{matrix} (inherits from \code{array}) with 18475 rows and 121 columns.
}
\source{
\url{https://cengen.org}
}
\usage{
cengen_TPM

cengen_proportion

cengen_UMI

cengen_nCells

cengen_TPM_bulk

cengen_proportion_bulk

cengen_UMI_bulk

cengen_nCells_bulk
}
\description{
Normalized expression of each gene in each neuron type of C. elegans by single-cell RNA-Seq.
}
\section{Details}{

A normalized value of the mRNA count of each gene in each neuron type. \emph{TPM} indicates a
normalization by individual cell and for each neuron type. \emph{proportion} indicates the
proportion of cells in each cluster where at least 1 UMI was detected. \emph{UMI} indicates the
total (non-normalized) number of UMIs detected in the cluster. Finally, \emph{nCells} gives the
total number of cells in each cluster (it is thus not a measure of expression level).

The dataset named `cengen_xx` indicates the expression levels are given with normalization `xx`.
The dataset named `cengen_xx_bulk` contains the same values as `cengen_xx`, but the columns names
have been changed to match those of bulk sequencing (see below).
}

\section{'Bulk' versions}{

The resolution of this single-cell dataset is not always consistent with annotation of other bulk
datasets, making comparisons more tedious. The `cengen_xx_bulk` are convenience objects where the
column names match the usual neuron names.

Specifically, for `cengen_TPM_bulk`, the single cells have been aggregated by (bulk) neuron type
then the TPM computed. For the other normalization methods, the normalized values have been averaged or
duplicated after normalization. Note that, while this usually doesn't lead to large changes, it is incorrect.
}

\section{Note}{

The thresholded tables available on the CeNGEN website contain this value of TPM
for the genes above the threshold.
}

\keyword{datasets}
