% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cengen_sc}
\alias{cengen_sc}
\alias{cengen_sc_1}
\alias{cengen_sc_2}
\alias{cengen_sc_3}
\alias{cengen_sc_4}
\alias{cengen_sc_1_bulk}
\alias{cengen_sc_2_bulk}
\alias{cengen_sc_3_bulk}
\alias{cengen_sc_4_bulk}
\title{CeNGEN single-cell thresholded expression.}
\format{
A matrix with one row per gene and one column by single-cell cluster:
\describe{
  \item{row name}{Wormbase ID of the gene}
  \item{column name}{cluster identification, usually corresponds to a neuron type}
  \item{value}{FALSE if the gene was below the threshold level, TRUE if it was above}
}

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 128 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 128 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 128 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 128 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 119 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 119 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 119 columns.

An object of class \code{matrix} (inherits from \code{array}) with 13669 rows and 119 columns.
}
\source{
\url{https://cengen.org}
}
\usage{
cengen_sc_1

cengen_sc_2

cengen_sc_3

cengen_sc_4

cengen_sc_1_bulk

cengen_sc_2_bulk

cengen_sc_3_bulk

cengen_sc_4_bulk
}
\description{
Binarized expression of each gene in each neuron type of C. elegans by single-cell RNA-Seq.
}
\section{Details}{

The proportion of cells from each cluster containing at
least 1 UMI for that gene was thresholded at several levels (from 1, the most
permissive, to 4, the most stringent). A value of FALSE indicates the gene is not expressed
in that cluster, a value of TRUE indicates it is expressed.

The dataset named `cengen_sc_n` indicates thresholding with threshold `n`. The dataset
named `cengen_sc_n_bulk` contains the same values as `cengen_sc_n`, but the columns names
have been changed to match those of bulk sequencing (some subtypes are not separate).
}

\keyword{datasets}
